// src/common/interfaces.ts
export interface RapidAPIResponse {
  uid?: string;
  mileage_adjustment?: number;
  loan_value?: number;
  trade_in_value?: number;
  adjusted_trade_in_value?: number;
  make?: string;
  make_code?: string;
  model?: string;
  model_code?: string;
  trim?: string;
  trim_code?: string;
  year?: number;
  retail_value?: number;
  msrp_value?: number;
  average_trade_in?: number;
  weight?: number;
}